requirements.txt
flask==3.0.0
python-docx==1.1.0
chromadb==0.5.18
sentence-transformers==3.3.1
requests>=2.31.0
python-dotenv==1.0.1